+++ packages/colors/src/colors.ts
@@ -1,0 +1,151 @@
+export const canaries = {
+  canary050: '#FFF1B9',
+  canary100: '#FFE784',
+  canary200: '#FDD778',
+  canary300: '#FFBE3C',
+  canary400: '#FFAC00',
+  canary500: '#FF9D00',
+  canary600: '#EF7600',
+  canary700: '#CE4D13',
+  canary800: '#833602',
+  canary900: '#4C2103',
+};
+
+export type CanaryNames = keyof typeof canaries;
+
+export const cockatoos = {
+  cockatoo025: '#FDD2FF',
+  cockatoo050: '#FFAEF9',
+  cockatoo100: '#FF83F3',
+  cockatoo200: '#FF4AE6',
+  cockatoo300: '#FF00D5',
+  cockatoo400: '#E004BC',
+  cockatoo500: '#B702A0',
+  cockatoo600: '#86017C',
+  cockatoo700: '#5D034E',
+  cockatoo800: '#3E0234',
+};
+
+export type CockatooNames = keyof typeof cockatoos;
+
+export const cuckoos = {
+  cuckoo100: '#EDE6FF',
+  cuckoo200: '#D7C6FF',
+  cuckoo300: '#C1A1FF',
+  cuckoo400: '#AC78FF',
+  cuckoo500: '#9F5CFF',
+  cuckoo600: '#9640FF',
+  cuckoo700: '#8029EF',
+  cuckoo800: '#5618B7',
+  cuckoo900: '#341078',
+};
+
+export type CuckooNames = keyof typeof cockatoos;
+
+export const flamingos = {
+  flamingo050: '#FCF0F0',
+  flamingo100: '#FFD9D9',
+  flamingo200: '#FF9797',
+  flamingo300: '#FF6969',
+  flamingo400: '#FF5050',
+  flamingo500: '#ED3B3B',
+  flamingo600: '#D23939',
+  flamingo700: '#A1222F',
+  flamingo800: '#841328',
+  flamingo900: '#511421',
+};
+
+export type FlamingoNames = keyof typeof flamingos;
+
+export const jays = {
+  jay020: '#E7F1FF',
+  jay050: '#DDE9FF',
+  jay100: '#B4CFFF',
+  jay150: '#90B8FF',
+  jay200: '#6FA2FF',
+  jay250: '#448AFF',
+  jay300: '#2E77FF',
+  jay350: '#2160FF',
+  jay400: '#034BFF',
+  jay450: '#052DE1',
+  jay500: '#0018CC',
+  jay600: '#0204A4',
+  jay700: '#000177',
+  jay800: '#010256',
+  jay900: '#020039',
+};
+
+export type JayNames = keyof typeof flamingos;
+
+export const macaws = {
+  macaw020: '#E7F1FF',
+  macaw040: '#DDE9FF',
+  macaw060: '#B4CFFF',
+  macaw080: '#90B8FF',
+  macaw100: '#6FA2FF',
+  macaw200: '#448AFF',
+  macaw300: '#2E77FF',
+  macaw350: '#2943FF',
+  macaw400: '#3000FF',
+  macaw500: '#2009D8',
+  macaw600: '#0F009E',
+};
+
+export type MacawNames = keyof typeof macaws;
+
+export const parrots = {
+  parrot025: '#DEFABB',
+  parrot050: '#C6F68D',
+  parrot100: '#AAF255',
+  parrot200: '#90EE02',
+  parrot300: '#75E900',
+  parrot400: '#61D800',
+  parrot500: '#41C300',
+  parrot600: '#09AF00',
+  parrot700: '#008B00',
+  parrot800: '#026802',
+};
+
+export type ParrotNames = keyof typeof parrots;
+
+export const peacocks = {
+  peacock050: '#E6F4F6',
+  peacock100: '#CEEDEF',
+  peacock200: '#B2ECED',
+  peacock300: '#70E8E8',
+  peacock400: '#33DBDB',
+  peacock500: '#07B3BE',
+  peacock600: '#05889F',
+  peacock700: '#036280',
+  peacock800: '#01486A',
+  peacock900: '#033B51',
+};
+export type PeacockNames = keyof typeof peacocks;
+
+export const pigeons = {
+  pigeon025: '#F7F7F7',
+  pigeon050: '#EFEFEF',
+  pigeon100: '#DCDCDC',
+  pigeon200: '#C1C1C1',
+  pigeon300: '#A4A4A4',
+  pigeon400: '#808080',
+  pigeon500: '#666666',
+  pigeon600: '#4D4D4D',
+  pigeon650: '#3F3E3E',
+  pigeon700: '#333333',
+  pigeon800: '#262626',
+  pigeon850: '#1A1A1A',
+  pigeon900: '#0F0F0F',
+};
+
+export type PigeonNames = keyof typeof pigeons;
+
+export const robins = {
+  robin300: '#FFAA3B',
+  robin400: '#FF9800',
+  robin500: '#FB8800',
+  robin600: '#FB7D00',
+  robin700: '#FF6D00',
+};
+
+export type RobinNames = keyof typeof robins;