+++ packages/zephyr/src/Modals/TransactionModal.tsx
@@ -7,7 +7,7 @@ export const TRANSACTION_MODAL_PRIMARY_CTA = 'TRANSACTION_MODAL_PRIMARY_CTA';
 export const TRANSACTION_MODAL_SECONDARY_CTA = 'TRANSACTION_MODAL_SECONDARY_CTA';
 export const TRANSACTION_MODAL_TERTIARY_CTA = 'TRANSACTION_MODAL_TERTIARY_CTA';
 
-export type TransactionModalButton = Omit<ButtonProps, 'ref'>;
+export type TransactionModalButton = Omit<ButtonProps, 'ref'> & { ['data-testid']?: string };
 
 export interface TransactionModalProps
   extends Omit<ModalProps, 'leastDestructiveRef' | 'withCloseButton' | 'isAlertModal'> {
@@ -29,8 +29,8 @@ export interface TransactionModalProps
   secondaryCTA?: TransactionModalButton | JSX.Element;
 
   /**
-   * This call-to-action (CTA) is optional and will appear in the bottom-left corner. If this CTA is not passed, nothing will be seen. You can pass any
-   * JSX.
+   * This call-to-action (CTA) is optional and will appear in the bottom-left corner. It renders a button ("button-ghost-destructive" variant)
+   * by default, but this can be overridden by passing JSX.
    */
   tertiaryCTA?: TransactionModalButton | JSX.Element;
 
@@ -74,6 +74,16 @@ export const TransactionModal = ({
     secondaryCTA
   );
 
+  const tertiaryCTAElement = !tertiaryCTA ? null : isUsingButtonSchema(tertiaryCTA) ? (
+    <Button
+      variant="button-ghost-destructive"
+      data-testid={TRANSACTION_MODAL_TERTIARY_CTA}
+      {...tertiaryCTA}
+    />
+  ) : (
+    tertiaryCTA
+  );
+
   return (
     <Modal
       className={className}
@@ -103,7 +113,7 @@ export const TransactionModal = ({
               alignItems: 'center',
             }}
           >
-            {tertiaryCTA}
+            {tertiaryCTAElement}
           </Box>
         )}
         <Box tx={{ '& > *:first-child': { mr: 12 } }}>