+++ packages/zephyr/src/RadixMenu/abstractions/RadixDropdownMenu.tsx
@@ -12,6 +12,7 @@ import { Box, BoxProps } from '../../Box';
 import { RadixMenuItem, RadixMenuItemProps } from '../components/RadixMenuItem';
 import { MenuSize } from '../../Menus/components/Menu';
 import { TXProp } from '../../theme';
+import { MenuVariantName } from '../../theme/variants/menus';
 
 export type RadixDropdownMenuOption = RadixMenuItemProps & {
   id?: string;
@@ -71,6 +72,8 @@ export interface RadixDropdownMenuProps extends Pick<MenuProps, 'animation' | 's
    * The trigger that will open the menu.
    */
   trigger: ReactNode;
+
+  variant?: MenuVariantName;
 }
 
 export const RadixDropdownMenu = memo(
@@ -83,6 +86,7 @@ export const RadixDropdownMenu = memo(
     options,
     onChange = noop,
     size = 'small',
+    variant = 'bright',
     ['data-testid']: testId,
     trigger,
     tx,
@@ -120,7 +124,7 @@ export const RadixDropdownMenu = memo(
     );
 
     return (
-      <Root onOpenChange={handleChange}>
+      <Root onOpenChange={handleChange} modal={false}>
         {/* This overlay is conditional rendered based on if `hasOverlay` prop is passed and it'll apply a transparent div over the entire screen to prevent the user from being able to right click on anything else while the menu is opened. */}
         {hasOverlay && (
           <Box
@@ -153,14 +157,14 @@ export const RadixDropdownMenu = memo(
                 display: 'flex',
                 flexDirection: 'column',
                 justifyContent: 'stretch',
-                backgroundColor: 'white',
+                backgroundColor: variant === 'dark' ? 'black' : 'white',
                 width: size === 'small' ? 216 : 240,
                 p: size === 'small' ? 6 : 8,
                 outline: 'none',
                 borderRadius: 4,
                 boxShadow: `
-                  0px 2px 8px ${rgba(theme.colors.black, 0.2)}, 
-                  0px 1px 3px ${rgba(theme.colors.black, 0.15)}, 
+                  0px 2px 8px ${rgba(theme.colors.black, 0.2)},
+                  0px 1px 3px ${rgba(theme.colors.black, 0.15)},
                   0px 0px 2px ${rgba(theme.colors.black, 0.25)}
                 `,
                 ...tx,
@@ -180,7 +184,8 @@ export const RadixDropdownMenu = memo(
                 return (
                   <RadixMenuItem
                     data-testid={option.id}
-                    onClick={(event: Event) => event.stopPropagation()}
+                    onClick={(event) => event.stopPropagation()}
+                    variant={variant}
                     key={index}
                     size={size}
                     {...option}