+++ packages/zephyr/src/RadixMenu/components/RadixMenuItem.tsx
@@ -4,19 +4,19 @@ import { Item, Root, Content, TriggerItem } from '@radix-ui/react-dropdown-menu'
 import { ChevronRight } from '@air/icons';
 import { useTheme } from 'styled-components';
 
-import { Box } from '../../Box';
+import { Box, BoxProps } from '../../Box';
 import { Text } from '../../Text';
 import { MenuItemDescription } from '../../Menus/components/MenuItemDescription';
 import { MenuItemLabel } from '../../Menus/components/MenuItemLabel';
 import { MenuSize } from '../../Menus/components/Menu';
-import { TXProp } from '../../theme';
+import { MenuVariantName } from '../../theme/variants/menus';
 import { RadixMenuItemDivider } from './RadixMenuItemDivider';
 
 export type RadixMenuItemRenderProps =
   | { children: ReactNode }
   | { label: ReactNode; description?: ReactNode };
 
-export type RadixMenuItemProps = {
+export type RadixMenuItemProps = Pick<BoxProps, 'tx' | 'id' | 'onClick'> & {
   /**
    * Renders `<MenuDivider />` below the menu item.
    */
@@ -52,6 +52,7 @@ export type RadixMenuItemProps = {
    * Calling event.preventDefault in this handler will prevent the dropdown menu from closing when selecting that item.
    */
   onClick?: (event: Event) => void;
+  onSelect?: () => void;
 
   /**
    * Prevents item selection.
@@ -59,9 +60,11 @@ export type RadixMenuItemProps = {
   disabled?: boolean;
 
   /**
-   * Allows the user to extend the items styles.
+   * Defaults to value of `id`
    */
-  tx?: TXProp;
+  ['data-testid']?: string;
+
+  variant?: MenuVariantName;
 
   subOptions?: (RadixMenuItemProps & RadixMenuItemRenderProps & { id?: string })[];
 } & RadixMenuItemRenderProps;
@@ -75,9 +78,13 @@ export const RadixMenuItem = memo(
     shortcut,
     size = 'small',
     onClick,
+    onSelect,
     tx,
     disabled,
     subOptions,
+    variant,
+    id,
+    'data-testid': testId,
     ...renderProps
   }: RadixMenuItemProps) => {
     const hasDescription = 'description' in renderProps;
@@ -87,32 +94,44 @@ export const RadixMenuItem = memo(
 
     const menuItemStyle = useMemo(
       () => ({
+        mb: hasDividerBottom ? 0 : 8,
+
+        '&:last-child': {
+          mb: 0,
+        },
+
         ['.radix-menu-item']: {
           display: 'flex',
           alignItems: hasDescription ? 'flex-start' : 'center',
           justifyContent: 'space-between',
           backgroundColor: 'transparent',
           height: hasDescription ? 'auto' : isSmallSize ? 32 : 36,
-          mb: hasDividerBottom ? 0 : 8,
           px: 6,
           py: hasDescription ? 6 : 0,
           border: 0,
           borderRadius: 4,
-          color: 'pigeon700',
+          color: variant === 'dark' ? 'pigeon100' : 'pigeon700',
           textAlign: 'left',
           cursor: 'pointer',
 
-          '&[data-selected]': { backgroundColor: 'pigeon050', color: 'pigeon700' },
+          '&[data-selected], &:active': {
+            backgroundColor: variant === 'dark' ? 'pigeon800' : 'pigeon050',
+            color: variant === 'dark' ? 'pigeon100' : 'pigeon700',
+          },
 
-          '&:hover': { backgroundColor: 'pigeon050' },
+          '&:hover': {
+            backgroundColor: variant === 'dark' ? 'pigeon800' : 'pigeon050',
+            color: variant === 'dark' ? 'white' : undefined,
+          },
 
-          '&:focus': { outline: 'none', backgroundColor: 'pigeon050' },
-
-          '&:last-child': { mb: 0 },
+          '&:focus': {
+            outline: 'none',
+            backgroundColor: variant === 'dark' ? 'pigeon700' : 'pigeon050',
+          },
           ...tx,
         },
       }),
-      [hasDescription, hasDividerBottom, isSmallSize, tx],
+      [hasDescription, hasDividerBottom, isSmallSize, tx, variant],
     );
 
     const menuItemContent = useMemo(
@@ -132,7 +151,10 @@ export const RadixMenuItem = memo(
                 <MenuItemLabel variant={isSmallSize ? 'text-ui-14' : 'text-ui-16'}>
                   {renderProps.label}
                 </MenuItemLabel>
-                <MenuItemDescription variant={isSmallSize ? 'text-ui-12' : 'text-ui-14'}>
+                <MenuItemDescription
+                  variant={isSmallSize ? 'text-ui-12' : 'text-ui-14'}
+                  tx={variant === 'dark' ? { color: 'pigeon100' } : undefined}
+                >
                   {renderProps.description}
                 </MenuItemDescription>
               </Box>
@@ -196,6 +218,7 @@ export const RadixMenuItem = memo(
         numberOfShortcutKeys,
         renderProps,
         rightAdornment,
+        variant,
         shortcut,
       ],
     );
@@ -215,14 +238,14 @@ export const RadixMenuItem = memo(
                   display: 'flex',
                   flexDirection: 'column',
                   justifyContent: 'stretch',
-                  backgroundColor: 'white',
+                  backgroundColor: variant === 'dark' ? 'black' : 'white',
                   width: size === 'small' ? 216 : 240,
                   p: size === 'small' ? 6 : 8,
                   outline: 'none',
                   borderRadius: 4,
                   boxShadow: `
-                    0px 2px 8px ${rgba(theme.colors.black, 0.2)}, 
-                    0px 1px 3px ${rgba(theme.colors.black, 0.15)}, 
+                    0px 2px 8px ${rgba(theme.colors.black, 0.2)},
+                    0px 1px 3px ${rgba(theme.colors.black, 0.15)},
                     0px 0px 2px ${rgba(theme.colors.black, 0.25)}
                   `,
                 }}
@@ -230,9 +253,10 @@ export const RadixMenuItem = memo(
                 {subOptions.map((option, index) => (
                   <RadixMenuItem
                     data-testid={option.id}
-                    onClick={(event: Event) => event.stopPropagation()}
+                    onClick={(event) => event.stopPropagation()}
                     key={index}
                     size={size}
+                    variant={variant}
                     {...option}
                   />
                 ))}
@@ -244,12 +268,18 @@ export const RadixMenuItem = memo(
     }
 
     return (
-      <Box tx={menuItemStyle}>
-        {hasDividerTop && <RadixMenuItemDivider isTop />}
-        <Item className="radix-menu-item" disabled={disabled} onSelect={onClick}>
+      <Box tx={menuItemStyle} {...renderProps}>
+        {hasDividerTop && <RadixMenuItemDivider variant={variant} isTop />}
+        <Item
+          data-testid={testId ?? id}
+          id={id}
+          className="radix-menu-item"
+          disabled={disabled}
+          onSelect={onClick ?? onSelect}
+        >
           {menuItemContent}
         </Item>
-        {hasDividerBottom && <RadixMenuItemDivider />}
+        {hasDividerBottom && <RadixMenuItemDivider variant={variant} />}
       </Box>
     );
   },