+++ packages/zephyr/stories/modals/TransactionModal.stories.tsx
@@ -76,7 +76,11 @@ export const TypicalTransactionModal: Story<TransactionModalProps> = () => {
             onDismiss={closeModal}
             modalLabel="Are you sure?"
             modalDescription="This is a transaction! We want to confirm that you want to do this, before you do it."
-            primaryCTA={{ children: `I'm sure!`, onClick: closeModal }}
+            primaryCTA={{
+              children: `I'm sure!`,
+              onClick: closeModal,
+              'data-testid': 'CUSTOM_PRIMARY_CTA_TEST_ID',
+            }}
             secondaryCTA={{ children: 'Nevermind', onClick: closeModal }}
           />
         )}
@@ -135,6 +139,16 @@ export const TertiaryTransactionModal: Story<TransactionModalProps> = () => {
   const openModal = () => setIsModalOpen(true);
   const closeModal = () => setIsModalOpen(false);
 
+  const defaultTertiaryCTA = {
+    children: 'Default CTA',
+    onClick: () => setIsChecked(true),
+    'data-testid': 'MY_CUSTOM_TERTIARY_TEST_ID',
+  };
+
+  const checkboxTertiaryCTA = (
+    <Checkbox label="Custom CTA Checkbox" checked={false} onChange={() => setIsChecked(false)} />
+  );
+
   return (
     <>
       <Button type="button" onClick={openModal} variant="button-filled-blue">
@@ -144,13 +158,7 @@ export const TertiaryTransactionModal: Story<TransactionModalProps> = () => {
       <AnimatePresence>
         {isModalOpen && (
           <TransactionModal
-            tertiaryCTA={
-              <Checkbox
-                label="Apply for all items"
-                checked={isChecked}
-                onChange={() => setIsChecked(!isChecked)}
-              />
-            }
+            tertiaryCTA={isChecked ? checkboxTertiaryCTA : defaultTertiaryCTA}
             onDismiss={closeModal}
             modalLabel="Uh oh! You have duplicates"
             modalDescription="3 out of 3 of the items you'd like to upload already exist. What would you like to do for item 1 of 3?"